coding

+++ goj-dal/src/main/java/com/gzu/taurus/goj/dal/dao/problem/SubmitDAO.java
@@ -22,7 +22,7 @@ public interface SubmitDAO {
 	 * @param submit
 	 * @return
 	 */
-	int createSubmit(SubmitDO submit);
+	Long createSubmit(SubmitDO submit);
 
 	/**
 	 * getSubmit
